Refactors check script
